refactor: extract horizon expanding into it's own file
